ConstrainedAspectRatio to size board nicely.
Kludge, but now if the board is tall we use all the
height, and if the board is wide we narrow the
entire gameboard.

This should let us put settings on the right
side and ensure we still look good.

Still looks jarring when we zoom in; need a
gentle animation when the longest stack
gets shorter.

Would also be great to do this without being
such a kludge.  We use layoutbuilder everywhere,
which ConstrainedAspectRatio was supposed
to solve.  Might as well just have the GameBoard
be an inheritedwidget that things read the size
on.  Or maybe make a Provider of card
size, which the GameBoard sets and notifies
all listeners, which will probably crash because
it's triggering layout during build or something.