Riverpod refactor! due to GlobalKey crazy bug:
Hours of debugging why cards were being rebuilt when they were about to
be kicked out of the widget tree.  Traced finally to the fact that the
card watched gameState (unnecessarily), and was in a
LayoutBuilder, so Riverpod didn't know after build phase and
before layout phase whether the card's widget instance
would exist after layout.  After build it still did, so it got rebuilt.
Which pissed it off since it had a GlobalKey, and there can be
only one of those.

Come to think of it, I didn't actually test whether LayoutBuilder
was the problem here and not Riverpod.  Maybe the GlobalKey
would still be in the tree twice.  Maybe it's not Riverpod's fault,
it just notified whoever was in the tree.

Anyway, as part of all this I pared down all the heavy
ref.watch(gameState)s that I had been doing, to only select
the bits needed.  Results in much less rebuilding.  Especially
all 52 cards were being rebuilt at every change (and probably
just about everything else as well).